app/pyproject.toml
[tool.poetry]
authors = ["FL03 <jo3mccain@icloud.com>"]
description = ""
name = "app"
packages = [{ include = "app", from = "src" }]
readme = "README.md"
version = "0.1.0"

[tool.poetry.dependencies]
bcrypt = "^4.0.0"
cryptography = "^38.0.1"
fastapi = { extras = ["all"], version = "^0.85.0" }
passlib = "^1.7.4"
python = ">=3.7,<3.11"
python-dotenv = "^0.21.0"
python-jose = "^3.3.0"
tortoise-orm = { extras = ["asyncpg"], version = "^0.19.2" }

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
